rename logging modules to be private
